Display yellow badge when user has a new exposure
Why:
As a user, I want to be alerted in the app when I have a new exposure so
that I know where to go to see the details of the exposure.

This commit:
Adds the ability to set a user's exposure status to the. When a new
exposure is detected, the exposure status is set to `true`, and when a
user visits the calendar tab, the exposure status will be set to
`false`. In both cases, the value will be persisted to AsyncStorage, so
that if they force quit the app, the badge will still be visible when
you re-launch.

Co-Authored-By: Alejandro Dustet <alejandro@thoughtbot.com>